fix(engine): warn on unknown reference fields
Misspelled reference fields (lang: for language:, parent: for
container:) were silently swallowed by the forward-compat
unknown_fields catch-all on every class payload, with no warning
anywhere in the pipeline. This caused undetected data loss in the
CNE Chicago fixtures of PR #895.

Add InputReference::unknown_fields() (class_dispatch accessor) and
unknown_reference_field_warnings(), emitted through the existing
central Warning aggregation in the session and format_document
paths so all adapters (CLI, WASM, FFI) surface it. citum check
reports the warnings and fails under --strict; citum render refs
prints them to stderr at authoring time.

Bean: csl26-q1nc
