+++ crates/citum-engine/src/api/mod.rs
@@ -15,11 +15,11 @@ mod refs_input;
 mod session;
 mod style_input;
 mod types;
+mod warnings;
 
 pub use document::{
     FormatDocumentError, FormatDocumentRequest, FormatDocumentResult, apply_style_overrides,
     format_document, format_document_with_resolver, format_document_with_style,
-    unknown_enum_warnings, unknown_reference_class_warnings,
 };
 pub use forward_compat::{UnknownFieldPath, collect_unknown_field_paths};
 pub use refs_input::RefsInput;
@@ -33,4 +33,7 @@ pub use types::{
     BibliographyEntry, CitationOccurrence, CitationOccurrenceItem, DocumentOptions, EntryMetadata,
     FormattedBibliography, FormattedBibliographyBlock, FormattedCitation, OutputFormatKind,
     Warning, WarningLevel,
+};
+pub use warnings::{
+    unknown_enum_warnings, unknown_reference_class_warnings, unknown_reference_field_warnings,
 };
